use idle timer hook
