+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "tidebit-defi",
-  "version": "0.9.0+6",
+  "version": "0.9.0+7",
   "private": true,
   "scripts": {
     "dev": "next dev",
