+++ tsconfig.json
@@ -2,7 +2,11 @@
   "compilerOptions": {
     "baseUrl": ".",
     "target": "ES2017",
-    "lib": ["dom", "dom.iterable", "esnext"],
+    "lib": [
+      "dom",
+      "dom.iterable",
+      "esnext"
+    ],
     "allowJs": true,
     "skipLibCheck": true,
     "strict": true,
@@ -12,7 +16,7 @@
     "moduleResolution": "bundler",
     "resolveJsonModule": true,
     "isolatedModules": true,
-    "jsx": "preserve",
+    "jsx": "react-jsx",
     "incremental": true,
     "plugins": [
       {
@@ -20,9 +24,15 @@
       }
     ],
     "paths": {
-      "@/*": ["./src/*"],
-      "@/package": ["./package.json"],
-      "@/i18nconfig": ["./i18n-config.ts"]
+      "@/*": [
+        "./src/*"
+      ],
+      "@/package": [
+        "./package.json"
+      ],
+      "@/i18nconfig": [
+        "./i18n-config.ts"
+      ]
     }
   },
   "include": [
@@ -32,7 +42,15 @@
     "src/**/*.tsx",
     "src/app/page.tsx",
     // "scripts/**/*.ts",
-    ".next/types/**/*.ts"
+    ".next/types/**/*.ts",
+    ".next/dev/types/**/*.ts"
   ],
-  "exclude": ["node_modules", ".next", "scripts", "test", "ignition", "dist"]
+  "exclude": [
+    "node_modules",
+    ".next",
+    "scripts",
+    "test",
+    "ignition",
+    "dist"
+  ]
 }
